UI: move search input to navbar
