+++ frontend/src/pages/search.js
@@ -1,18 +1,14 @@
-import React, { useContext, useEffect, useRef } from 'react'
+import React, { useEffect, useRef } from 'react'
 import Link from 'next/link'
 import { arrayOf, object, string } from 'prop-types'
 
-import { Input, Form, Loader } from 'semantic-ui-react'
+import { Loader } from 'semantic-ui-react'
 
 import Page from '@components/Page'
-import useForm from '@hooks/useForm'
 import { useSearchRepos } from '@hooks/Gitea'
-import { AuthContext } from '@contexts/AuthContext'
-import SearchProvider, { useSearchQuery } from '@contexts/SearchContext'
+import { useSearchQuery } from '@contexts/SearchContext'
 import { getAllRepos, searchRepos } from '@utils/giteaApi'
-import SearchFormModel from '@models/SearchFrom'
 import ProjectCard from '@components/ProjectCard'
-import { useRouter } from 'next/router'
 import { getFlatProjects } from '@utils/projectPage'
 
 import styles from './search.module.scss'
@@ -35,17 +31,9 @@ export const getServerSideProps = async ({ query }) => {
 }
 
 const Search = ({ initialProjects, initialQuery }) => {
-  const { user } = useContext(AuthContext)
-
-  const username = user?.login ?? 'unknown user'
-
   return (
-    <Page title="Kitspace | Home">
-      <div>Hi there {username}</div>
-      <SearchProvider initialQuery={initialQuery}>
-        <SearchForm />
-        <CardsGrid initialProjects={initialProjects} />
-      </SearchProvider>
+    <Page initialQuery={initialQuery} title="Kitspace | Home">
+      <CardsGrid initialProjects={initialProjects} />
     </Page>
   )
 }
@@ -91,46 +79,6 @@ const CardsGrid = ({ initialProjects }) => {
       {projects?.map(project => (
         <ProjectCard {...project} key={project.id} />
       ))}
-    </div>
-  )
-}
-
-const SearchForm = () => {
-  const { form, onChange, formatErrorPrompt, populate } = useForm(SearchFormModel)
-  const { push } = useRouter()
-  const { updateQuery, query } = useSearchQuery()
-
-  useEffect(() => {
-    populate({ query })
-  }, [populate, query])
-
-  /**
-   * i.   Update the search query in the {@link SearchProvider}.
-   * ii.  Change the search query parameter `q`.
-   * ii.  Make a `shallow` redirect to the new url `/search?q=${submitted query term}`.
-   */
-  const onSubmit = () => {
-    updateQuery(form.query)
-    push(`/search?q=${form.query}`, undefined, { shallow: true })
-  }
-
-  return (
-    <div className={styles.searchForm}>
-      <Form onSubmit={onSubmit}>
-        <Form.Group widths="equal" className={styles.searchFormGroup}>
-          <Form.Field
-            data-cy="search-field"
-            icon="search"
-            fluid
-            control={Input}
-            placeholder="Search for projects"
-            name="query"
-            value={form.query ?? ''}
-            onChange={onChange}
-            error={form.query !== '' && formatErrorPrompt('query')}
-          />
-        </Form.Group>
-      </Form>
     </div>
   )
 }
